app/requirements.txt
fastapi
uvicorn
requests
pillow
pytesseract
numpy
markdown
speechrecognition
beautifulsoup4
